feat: add test for form
